tests/wpt/web-platform-tests/mediacapture-streams/parallel-capture-requests.https.html

<!doctype html>
<html>
<head>
<title>Parallel capture requests</title>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(function() {
  const getUserMediaPromise =
        navigator.mediaDevices.getUserMedia({audio: true, video:true});
  const getDisplayMediaPromise =
        navigator.mediaDevices.getDisplayMedia({video: true, audio: true});
  return Promise.all([getUserMediaPromise, getDisplayMediaPromise])
      .then(function(s) {
        assert_equals(s[0].getTracks().length, 2);
        assert_equals(s[0].getAudioTracks().length, 1);
        assert_equals(s[0].getVideoTracks().length, 1);
        assert_equals(s[1].getTracks().length, 2);
        assert_equals(s[1].getAudioTracks().length, 1);
        assert_equals(s[1].getVideoTracks().length, 1);
      });
}, 'getDisplayMedia() and parallel getUserMedia()');

promise_test(function() {
  const getDisplayMediaPromise =
        navigator.mediaDevices.getDisplayMedia({video: true, audio: true});
  const getUserMediaPromise =
        navigator.mediaDevices.getUserMedia({audio: true, video:true});
  return Promise.all([getDisplayMediaPromise, getUserMediaPromise])
      .then(function(s) {
        assert_equals(s[0].getTracks().length, 2);
        assert_equals(s[0].getAudioTracks().length, 1);
        assert_equals(s[0].getVideoTracks().length, 1);
        assert_equals(s[1].getTracks().length, 2);
        assert_equals(s[1].getAudioTracks().length, 1);
        assert_equals(s[1].getVideoTracks().length, 1);
      });
}, 'getUserMedia() and parallel getDisplayMedia()');
</script>
</body>
</html>
